// ui/src/phaser3-rex-plugins/templates/ui/ui-plugin.d.ts
declare class Factories {
    alphaMaskImage: typeof AlphaMaskImageFactory;
    anchor: typeof AnchorFactory;
    badgeLabel: typeof BadgeLabelFactory;
    BBCodeText: typeof BBCodeTextFactory;
    buttons: typeof ButtonsFactory;
    canvas: typeof CanvasFactory;
    canvasInput: typeof CanvasInputFactory;
    chart: typeof ChartFactory;
    checkbox: typeof CheckboxFactory;
    circleMaskImage: typeof CircleMaskImageFactory;
    circularProgressCanvas: typeof CircularProgressCanvasFactory;
    circularProgress: typeof CircularProgressFactory;
    click: typeof ClickFactory;
    clickOutside: typeof ClickOutsideFactory;
    colorComponents: typeof ColorComponentsFactory;
    colorInput: typeof ColorInputFactory;
    colorInputLite: typeof ColorInputLiteFactory;
    colorPicker: typeof ColorPickerFactory;
    confirmDialog: typeof ConfirmDialogFactory;
    inTouching: typeof InTouchingFactory;
    container: typeof ContainerFactory;
    cover: typeof CoverFactory;
    customProgress: typeof CustomProgressFactory;
    customShapes: typeof CustomShapesFactory;
    dialog: typeof DialogFactory;
    drag: typeof DragFactory;
    dropDownList: typeof DropDownListFactory;
    dynamicTextFactory: typeof DynamicTextFactory;
    flip: typeof FlipFactory;
    fileChooser: typeof FileChooserFactory;
    fileDropZone: typeof FileDropZoneFactory;
    fileSelectorButton: typeof FileSelectorButtonFactory;
    fixWidthButtons: typeof FixWidthButtonsFactory;
    fixWidthSizer: typeof FixWidthSizerFactory;
    folder: typeof FolderFactory;
    fullWindowRectangle: typeof FullWindowRectangleFactory;
    gridButtons: typeof GridButtonsFactory;
    gridSizer: typeof GridSizerFactory;
    gridTable: typeof GridTableFactory;
    hiddenEdit: typeof HiddenEditFactory;
    holyGrail: typeof HolyGrailFactory;
    knob: typeof KnobFactory;
    imageBox: typeof ImageBoxFactory;
    inputText: typeof InputTextFactory;
    label: typeof LabelFactory;
    lineProgress: typeof LineProgressFactory;
    lineProgressCanvas: typeof LineProgressCanvasFactory;
    // maker: typeof MakerFactory;
    menu: typeof MenuFactory;
    nameValueLabel: typeof NameValueLabelFactory;
    ninePatch: typeof NinePatchFactory;
    ninePatch2: typeof NinePatch2Factory;
    numberBar: typeof NumberBarFactory;
    overlapSizer: typeof OverlapSizerFactory;
    pages: typeof PagesFactory;
    pan: typeof PanFactory;
    perspective: typeof PerspectiveFactory;
    perspectiveCard: typeof PerspectiveCardFactory;
    pinch: typeof PinchFactory;
    press: typeof PressFactory;
    rotate: typeof RotateFactory;
    roundRectangleCanvas: typeof RoundRectangleCanvasFactory;
    roundRectangle: typeof RoundRectangleFactory;
    scrollablePanel: typeof ScrollablePanelFactory;
    scrollBar: typeof ScrollBarFactory;
    simpleLabel: typeof SimpleLabelFactory;
    simpleDropDownList: typeof SimpleDropDownListFactory;
    sizer: typeof SizerFactory;
    shake: typeof ShakeFactory;
    skew: typeof SkewFactory;
    slider: typeof SliderFactory;
    space: typeof SpaceFactory;
    statesRoundRectangle: typeof StatesRoundRectangleFactory;
    swipe: typeof SwipeFactory;
    tabs: typeof TabsFactory;
    tabPages: typeof TabPagesFactory;
    tagText: typeof TagTextFactory;
    tap: typeof TapFactory;
    textArea: typeof TextAreaFactory;
    textBox: typeof TextBoxFactory;
    textEdit: typeof TextEditFactory;
    textPage: typeof TextPageFactory;
    textPlayer: typeof TextPlayerFactory;
    textTyping: typeof TextTypingFactory;
    toast: typeof ToastFactory;
    toggleSwitch: typeof ToggleSwitchFactory;
    touchEventStop: typeof TouchEventStopFactory;
    triangle: typeof TriangleFactory;
    tweaker: typeof TweakerFactory;
    twoRowsLabel: typeof TitleLabelFactory;
}

declare class UIPlugin extends Phaser.Plugins.ScenePlugin {
    add: Factories;


    easeMoveTo: typeof EaseMoveTo;
    easeMoveFrom: typeof EaseMoveFrom;
    edit: typeof Edit;
    fadeIn: typeof FadeIn;
    fadeOutDestroy: typeof FadeOutDestroy;
    modal: typeof Modal;
    modalPromise: typeof ModalPromise;
    modalClose: typeof ModalClose;
    getParentSizer: typeof GetParentSizer;
    getTopmostSizer: typeof GetTopmostSizer;
    hide: typeof Hide;
    show: typeof Show;
    isShown: typeof IsShown;
    setChildrenInteractive: typeof SetChildrenInteractive;
    waitEvent: typeof WaitEvent;
    waitComplete: typeof WaitComplete;
    delayPromise: typeof Delay;
    wrapExpandText: typeof WrapExpandText;
    fontSizeExpandText: typeof FontSizeExpandText;
    fontSizeResize: typeof SetFontSizeToFitWidth;  // Backward compatibility
    setFontSizeToFitWidth: typeof SetFontSizeToFitWidth;
    requestDrag: typeof RequestDrag;
    openFileChooser: typeof OpenFileChooser;

    isInTouching(
        gameObject: Phaser.GameObjects.GameObject,
        pointer?: Phaser.Input.Pointer,
        preTest?: (gameObject: Phaser.GameObjects.GameObject, x: number, y: number) => boolean,
        postTest?: (gameObject: Phaser.GameObjects.GameObject, x: number, y: number) => boolean,
    ): boolean;

    readonly viewport: Phaser.Geom.Rectangle;
}
// ...
declare namespace UIPlugin {
    type AlphaMaskImage = AlphaMaskImageClass;
    type BadgeLabel = BadgeLabelClass;
    type BBCodeText = BBCodeTextClass;
    type Buttons = ButtonsClass;
    type Canvas = CanvasClass;
    type CanvasInput = CanvasInputClass;
    type Checkbox = CheckboxClass;
    type CircleMaskImage = CircleMaskImageClass;
    type CircularProgressCanvas = CircularProgressCanvasClass;
    type CircularProgress = CircularProgressClass;
    type Click = ClickClass;
    type ClickOutside = ClickOutsideClass;
    type ColorComponents = ColorComponentsClass;
    type ColorInput = ColorInputClass;
    type ColorInputBase = ColorInputLiteClass;
    type ColorPicker = ColorPickerClass;
    type Container = ContainerClass;
    type ConfirmDialog = ConfirmDialogClass;
    type Cover = CoverClass;
    type CustomProgress = CustomProgressClass;
    type CustomShapes = CustomShapesClass;
    type Dialog = DialogClass;
    type Drag = DragClass;
    type DropDownList = DropDownListClass;
    type DynamicText = DynamicTextClass;
    type EaseMove = EaseMoveClass;
    type Fade = FadeClass;
    type Flip = FlipClass;
    type FileChooser = FileChooserClass;
    type FileDropZone = FileDropZoneClass;
    type FileSelectorButton = FileSelectorButtonClass;
    type FixWidthButtons = FixWidthButtonsClass;
    type Folder = FolderClass;
    type FullWindowRectangle = FullWindowRectangleClass;
    type FixWidthSizer = FixWidthSizerClass;
    type GridButtons = GridButtonsClass;
    type GridSizer = GridSizerClass;
    type GridTable = GridTableClass;
    type HiddenEdit = HiddenTextEditClass;
    type HolyGrail = HolyGrailClass;
    type ImageBox = ImageBoxClass;
    type InputText = InputTextClass;
    type InTouching = InTouchingClass;
    type Knob = KnobClass;
    type Label = LabelClass;
    type LineProgress = LineProgressClass;
    type LineProgressCanvas = LineProgressCanvasClass;
    // type maker = MakerClass;
    type Menu = MenuClass;
    type NameValueLabel = NameValueLabelClass;
    type NinePatch = NinePatchClass;
    type NumberBar = NumberBarClass;
    type OverlapSizer = OverlapSizerClass;
    type Pages = PagesClass;
    type Pan = PanClass;
    type Perspective = PerspectiveClass;
    type PerspectiveCard = PerspectiveCardClass;
    type Pinch = PinchClass;
    type Press = PressClass;
    type Rotate = RotateClass;
    type RoundRectangleCanvas = RoundRectangleCanvasClass;
    type RoundRectangle = RoundRectangleClass;
    type ScrollablePanel = ScrollablePanelClass;
    type ScrollBar = ScrollBarClass;
    type SimpleDropDownList = SimpleDropDownListClass;
    type SimpleLabel = SimpleLabelClass;
    type Sizer = SizerClass;
    type shake = ShakeClass;
    type skew = SkewClass;
    type Slider = SliderClass;
    type Swipe = SwipeClass;
    type Space = SpaceClass;
    type StatesRoundRectangle = StatesRoundRectangleClass;
    type Tabs = TabsClass;
    type TabPages = TabPagesClass;
    type TagText = TagTextClass;
    type Tap = TapClass;
    type TextArea = TextAreaClass;
    type TextBox = TextBoxClass;
    type TextPage = TextPageClass;
    type TextPlayer = TextPlayerClass;
    type TextTyping = TextTypingClass;
    type TitleLabel = TitleLabelClass;
    type Toast = ToastClass;
    type ToggleSwitch = ToggleSwitchClass;
    type TouchEventStop = TouchEventStopClass;
    type Triangle = TriangleClass;
    type Tweaker = TweakerClass;
}